moved spacy utils
they are indev and used for prettification
